Add process_blank.c to refactor code
