+++ src/lexer/lexer_internal.h
@@ -6,12 +6,14 @@
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/30 22:25:14 by yliu              #+#    #+#             */
-/*   Updated: 2024/08/30 22:36:27 by yliu             ###   ########.fr       */
+/*   Updated: 2024/08/30 22:42:36 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LEXER_INTERNAL_H
 # define LEXER_INTERNAL_H
+
+#include "token.h"
 
 typedef struct s_lexer {
   char *start;
@@ -19,5 +21,6 @@ typedef struct s_lexer {
 } t_lexer;
 
 void construct_lexer(char *string, t_lexer *lexer);
+t_token_list *process_blank(t_lexer *lexer);
 
 #endif
